Separate regex.cc into header and implementation. Remove various functions
