+++ src/C1402_regex.cc
@@ -3,131 +3,9 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
-#include <cctype>
-#include <stdexcept>
-#include <string.h>
-#include <iostream>
-#include <string>
-#include <catch2/catch_test_macros.hpp>
-/* 1."abc" -> 	acceptsthe literal string "abc"
- * 2."a?bc" ->	accepts "abc" or "bc".
- * 3."ab*c" -> 	accepts "ac", "abc", "abbbbbbbbbbbbbbbbc"
- * 4."ab+c" -> 	accepts "abc", "abbbbbbbbbbbbbbc"
- * 5."ab(CD)" -> 	accepts "abC", "abD"
- * 6."ab(C*D*)"->	accepts "ab", "abC", "abD", "abCCCCCCCC", "abDDDDDD"
- * 7."ab(CD)*" -> acceptance to be determined when compilation/parsing code is rewritten. (Possibly make equivalent to input string 6? Pre-processing or somesuch method)
- * 8. "ab." -> 	accepts "abZ", "abf", "ab9"
- * 9. "ab&" -> accepts "aba", "abB", "abz"
- * 10."ab#" -> accepts "ab1", "ab2", "ab9"
- * 11. "ab.*" -> accepts "ab", "ab1", "ab22222222222", "abf", "abZZZZZZZZz"
- * */
-#ifndef PRINT_MESSAGES
-#define PRINT_MESSAGES 1
-#endif
+#include <C1402_regex.h>
+using namespace C1402_regex;
 //TODO: Implemented longest-match functionality for repetition options.
-enum rules {
-	R_DEFAULT,
-	R_STAR,
-	R_PLUS,
-	R_OPT,
-};
-enum substitution_type {
-	S_LITERAL,
-	S_DIGIT, 	//DIGIT - numbers 0-9
-	S_ALPHA, 	//ALPHA - letters a-zA-Z
-	S_ALNUM, //ALPHANUMERIC - any letter or number
-};
-static std::string rule_to_string(rules rule) {
-	switch(rule) {
-		default:
-#if PRINT_MESSAGES==1
-			std::cout << "defaulting in rule_to_string";
-#endif
-			[[fallthrough]];
-		case(R_DEFAULT):
-			return "DEFAULT";
-		case(R_OPT):
-			return "OPTION";
-		case(R_PLUS): //visit notes and recall what the name of this is. kleene closure or something
-			return "PLUS";
-		case(R_STAR):
-			return "STAR";
-	}
-}
-std::string sub_to_string(substitution_type sub_rule) {
-	switch (sub_rule) {
-		case(S_ALNUM):
-			return "ALNUM";
-		case(S_ALPHA):
-			return "ALPHA";
-		case(S_DIGIT):
-			return "DIGIT";
-		default:
-			return "ERROR";
-		case(S_LITERAL):
-			return "LITERAL";		
-	}
-};
-/***                           	  the regex class                                   ***/
-class regex {
-	public:
-		regex(int, substitution_type);
-		virtual ~regex() {
-			if (alternate != NULL) {
-				delete alternate;
-			}
-			//because the alternate will also point to next
-			else if (next != NULL) {
-				delete next;
-			}
-		}
-		int getLiteral();
-
-		regex* getNext();
-		virtual void setNext(regex*);
-		//return a char* to the pointer where the match ends. This can be used to determine where the match occurred, and its length
-		virtual char* match_here(char* text);
-		virtual rules getRule() {
-			return R_DEFAULT;
-		}
-		bool accepts(int character);
-		void addAlternate(regex*);
-		regex* getAlternate();
-		//
-		std::string sub_as_string() {
-			return sub_to_string(sub_rule);
-		}
-	protected:
-		class regex* next;
-		class regex* alternate;
-		substitution_type sub_rule;
-	private:
-		//literal's value will not matter when substitute != R_CHAR
-		int literal;
-};
-class regex_star : public regex {
-	public:
-		regex_star(int _literal, substitution_type _sub_rule = S_LITERAL) :
-			regex(_literal,_sub_rule) {};
-		char* match_here(char* text) override;
-		rules getRule() override { return R_STAR; };
-};
-class regex_plus : public regex {
-	public:
-		regex_plus(int _literal, substitution_type _sub_rule = S_LITERAL) :
-			regex(_literal, _sub_rule) {};
-		char* match_here(char* text) override;
-		rules getRule() override { return R_PLUS; };
-};
-class regex_opt : public regex {
-	public:
-		regex_opt(int _literal, substitution_type _sub_rule = S_LITERAL) :
-			regex(_literal, _sub_rule) {};
-		char* match_here(char* text) override;
-		rules getRule() override { return R_OPT; };
-};
-/** 	      prototypes 	        **/
-void re_print(regex*);
 //creates a regex object with specified rule & literal
 regex::regex(int _literal, substitution_type _sub_rule = S_LITERAL) {
 	literal = _literal;
@@ -144,9 +22,6 @@ bool regex::accepts(int character) {
 		case(S_DIGIT):
 			return std::isdigit(character);
 		default:
-#if PRINT_MESSAGES==1
-			std::cout << "::accepts() - sub rule defaulting to S_LITERAL\n\n";
-#endif
 			[[fallthrough]];
 		case(S_LITERAL):
 			return character == literal;
@@ -176,16 +51,6 @@ void regex::addAlternate(regex* _alternate) {
 	else
 		alternate = _alternate;
 }
-//prints the nodes information
-void re_print(regex* instance) {
-	if (instance)
-		printf("[Rule: %s | Literal: %c | Next? : %s | Alt?: %s]",
-				rule_to_string(instance->getRule()).c_str(),
-				instance->getLiteral(),
-				(instance->getNext()) ? "Yes" : "No",
-				(instance->getAlternate() ? "yes" : "no"));
-	else printf("[-]");
-}
 /*** 					matching 				    ***/
 //returns true if the expression found a match in the text, otherwise it returns false.
 bool match(regex* expression, std::string input_text) {
@@ -194,9 +59,6 @@ bool match(regex* expression, std::string input_text) {
 	strncpy(text, input_text.c_str(), input_text.size());
 	char* end_of_match = NULL;
 	do {
-#if PRINT_MESSAGES==1
-			std::cout << "\n<-match-loop->\t" << text << "\n";
-#endif
 		end_of_match = expression->match_here(text);
 		if (end_of_match != NULL)
 			break;
@@ -213,15 +75,6 @@ bool match(regex* expression, std::string input_text) {
 	free(start_of_text);
 }
 char* regex::match_here(char *text) {
-#if PRINT_MESSAGES==1
-		std::cout << "match:\t";
-		if (sub_rule == S_LITERAL)
-			std::cout << (char)getLiteral() << " == " << *text;
-		else
-			std::cout << sub_as_string() << " == " << *text;
-		std::cout << "\t|" << "alt:" << std::string((alternate)?"T":"F") << "\t";
-		std::cout << "\t|" << "next:" << std::string((next)?"T":"F") << "\n";
-#endif
 /* Match
  * process current rule.
  * If returns true: process next.
@@ -241,14 +94,6 @@ char* regex::match_here(char *text) {
 	return NULL;
 }
 char* regex_star::match_here(char *text) {
-#if PRINT_MESSAGES==1
-		std::cout << "match*:\t";
-		if (sub_rule == S_LITERAL)
-			std::cout << (char)getLiteral() << " == " << *text;
-		else
-			std::cout << sub_as_string() << " == " << *text;
-		std::cout << "\t|" << "next:" << std::string((next)?"T":"F") << "\n";
-#endif
 	//perform the absolute shortest match if we have no options.
 	if (next == NULL) {
 		//the shortest possible match in R_STAR is NO match
@@ -256,7 +101,7 @@ char* regex_star::match_here(char *text) {
 	}
 	//continue attempting the current rule
 	else {
-		char* tmp_text = text; //used so that if/when we call the alternate case, we have the original starting point
+		char* tmp_text = text;
 		do {
 			auto retVal = next->match_here(tmp_text);
 			if(retVal != NULL) return retVal;
@@ -272,19 +117,9 @@ char* regex_star::match_here(char *text) {
 }
 
 char* regex_plus::match_here(char* text) {
-#if PRINT_MESSAGES==1
-		std::cout << "match+:\t";
-		if (sub_rule == S_LITERAL)
-			std::cout << (char)getLiteral() << " == " << *text;
-		else
-			std::cout << sub_as_string() << " == " << *text;
-		std::cout << "\t|" << "next:" << std::string((next)?"T":"F") << "\n";
-#endif
 	if (next == NULL && accepts(*text)) return text+1;
 	if (next != NULL) {
 		for (char* tmp_text = text;*tmp_text != '\0' && accepts(*tmp_text);tmp_text++) {
-			//we use the tmp array here so that if it fails we can attempt an alternate match.
-			//Otherwise, we could just store the offset instead of duplicating text
 			auto retVal = next->match_here(tmp_text+1);
 			if (retVal != NULL) return retVal;
 		}
@@ -296,14 +131,6 @@ char* regex_plus::match_here(char* text) {
 	return NULL;
 }
 char* regex_opt::match_here(char *text) {
-#if PRINT_MESSAGES==1
-		std::cout << "match?:\t";
-		if (sub_rule == S_LITERAL)
-			std::cout << (char)getLiteral() << " == " << *text;
-		else
-			std::cout << sub_as_string() << " == " << *text;
-		std::cout << "\t|" << "next:" << std::string((next)?"T":"F") << "\n";
-#endif
 	if (accepts(*text)) {
 		if (next != NULL ) return next->match_here(++text);
 		else return text+1;
@@ -320,11 +147,8 @@ char* regex_opt::match_here(char *text) {
 }
 
 /*** 				regex compilation 				    ***/
-/* The following code is for compiling the    *//** 	      prototypes 	        **/
-/*regex from an input string into the linked  */rules symbol_to_rrule(regex*, char);
-/*regex nodes.                                */substitution_type symbol_to_srule(char);
 //Return the rule associated with a symbol. R_DEFAULT if there is no matching rule
-rules symbol_to_rrule(char c) {
+rules C1402_regex::utils::symbol_to_rrule(char c) {
 	switch(c) {
 		case('*'):
 			return R_STAR;
@@ -337,7 +161,7 @@ rules symbol_to_rrule(char c) {
 	}
 }
 //Return the substitution rule associated with the symbol. S_LITERAL for non-matching symbol
-substitution_type symbol_to_srule(char c) {
+substitution_type C1402_regex::utils::symbol_to_srule(char c) {
 	switch(c) {
 		case('.'):
 			return S_ALNUM;
@@ -352,16 +176,13 @@ substitution_type symbol_to_srule(char c) {
 //TODO lint input, e.g., (ab)* is invalid, but (a*b*) is valid
 //TODO implement this as a constructor of regex
 //create a node-tree of REs from an input string
-regex* create_from_string(std::string regex_tape) {
+regex* C1402_regex::utils::create_from_string(std::string regex_tape) {
 	regex* root_node = NULL;
 	regex* last_node = NULL;
 	regex* alternate_node = NULL; //holds the node which has alternates
 	bool do_alternative = false;
 	bool escaped_symbol = false;
 	for (size_t i = 0; i < regex_tape.size(); i++) {
-#if PRINT_MESSAGES==1
-		std::cout << "cfs: i=" << i << "\n";
-#endif
 	//begin compilation
 		regex* current_node;
 		rules current_rule = R_DEFAULT;
@@ -370,39 +191,20 @@ regex* create_from_string(std::string regex_tape) {
 		auto subrule = symbol_to_srule(regex_tape.at(i));
 		int literal = regex_tape.at(i);
 		if (literal == '(') {
-#if PRINT_MESSAGES==1
-			std::cout << "set do_alternative. Skip...\n";
-#endif
 			do_alternative = true;
 			continue;
 		}
 		else if (literal == ')') {
-#if PRINT_MESSAGES==1
-			std::cout << "unset do_alternative. Skip...\n";
-#endif
 			do_alternative = false;
 			continue;
 		}
 		else if (literal == '\\' and escaped_symbol == false) {
-#if PRINT_MESSAGES==1
-			std::cout << "set escaped_symbol. Skip...\n";
-#endif
 			escaped_symbol = true;
 			continue;
 		}
-#if PRINT_MESSAGES==1
-		std::cout << "CR:\t" << rule_to_string(current_rule) << "\n";
-		std::cout << "SR:\t" << sub_to_string(subrule) << "\n";
-		std::cout << "L:\t" << (char)literal << "\n";
-		std::cout << "Alt?\t" << std::string((do_alternative)?"true":"false") << "\n";
-		std::cout << "Esc?\t" << std::string((escaped_symbol)?"true":"false") << "\n";
-#endif
 		if (escaped_symbol) {
 			subrule = S_LITERAL;
 			escaped_symbol = false;
-#if PRINT_MESSAGES==1
-			std::cout << "subRule = S_LITERAL, unset Esc\n";
-#endif
 		}
 		switch (current_rule) {
 			case R_DEFAULT:
@@ -442,21 +244,25 @@ regex* create_from_string(std::string regex_tape) {
 			last_node = current_node;
 		}
 		if (current_rule != R_DEFAULT) i+=1; //consume an extra character if we set a rule
-#if PRINT_MESSAGES==1
-		std::cout << "------\n";
-#endif
 	}//end compilation
 	if (alternate_node != NULL) {
-#if PRINT_MESSAGES==1
-		std::cout << "ending with alt node != NULL. last.setNext(alt)\n";
-#endif
 		//This occurs when the regex string ends with a ')'
 		last_node->setNext(alternate_node);
 	}
 	return root_node;	
 }
-#ifndef UNIT_TESTING
+#ifdef REGEX_MAIN
 /*** 				    main 					    ***/
+//prints the nodes information
+void re_print(regex* instance) {
+	if (instance)
+		printf("[Rule: %s | Literal: %c | Next? : %s | Alt?: %s]",
+				utils::rule_to_string(instance->getRule()).c_str(),
+				instance->getLiteral(),
+				(instance->getNext()) ? "Yes" : "No",
+				(instance->getAlternate() ? "yes" : "no"));
+	else printf("[-]");
+}
 int main(int argc, char** argv) {
 	int a; 				//iterator
 	int input_len = 0; 			//the length of the input string, for malloc
@@ -503,7 +309,8 @@ int main(int argc, char** argv) {
 	free(input_text);
 	delete (regexpr);
 }
-#else
+#endif
+#ifdef UNIT_TESTING
 #define CATCH_CONFIG_MAIN
 TEST_CASE( "test match on a single literal" ) {
 	regex* expression = create_from_string("a");
